% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_catch.R
\docType{data}
\name{fish_catch}
\alias{fish_catch}
\title{fish_catch}
\format{
A data frame with 300 rows and 1 variable
}
\source{
Dr. Naomi Tague & Will Burke
\itemize{
  \item fish. type of fish caught
}
}
\usage{
data(fish_catch)
}
\description{
Contains records of 300 fish caught
}
\keyword{datasets}
