% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_count.R
\name{fish_count}
\alias{fish_count}
\title{fish_count}
\usage{
fish_count(fish, hist.plot = FALSE)
}
\arguments{
\item{hist.plot}{}
}
\value{
Output will be a list.  If hist.plot = TRUE, will return a histogram
}
\description{
Returns a list of the most fish caught, rarest fish caught, and total fish caught.  If hist.plot = TRUE, returns a histogram of fish counts.
}
