% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_50.R
\name{ld_50}
\alias{ld_50}
\title{ld_50}
\usage{
ld_50(lb_bodywt, cups = 0, mg_caf = 95, caf_lvl = 0)
}
\arguments{
\item{caf_lvl}{}
}
\value{
Output is a numeric.
}
\description{
Returns the amount of cups of coffee it would take to reach a lethal dose of caffeine overdose given input weight in lbs.  Based on the LD 50 of caffeine (150 mg/kg body weight).
}
