% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_speed.R
\name{safe_speed}
\alias{safe_speed}
\title{Safe Speed}
\usage{
safe_speed(v, mu, safe_distance, g = 9.8, d = 0)
}
\arguments{
\item{v}{The starting speed for calculation. After each attempt, the velocity will increase 0.1 unit. (m/s)}

\item{mu}{The friction coefficient, need to identify for different roads.}

\item{safe_distance}{The goal of braking distance. (m)}

\item{g}{The acceleration of gravity, a constant. (9.8 m/s)}

\item{d}{The technical braking distance to be calculate in the loop, the initial default is 0. (m)}
}
\value{
The highest speed estimate until the target braking distance is reached. Numeric.
}
\description{
Estimate the highest "velocity" to stop a car for a given distance for different road (with different friction coefficient)
}
\author{
Ruoyu Wang
}
