% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years_to_K.R
\name{years_to_K}
\alias{years_to_K}
\title{Calculate the number of years until a population reaches carrying capacity}
\usage{
years_to_K(starting_stock, r, K)
}
\arguments{
\item{starting_stock}{The starting stock of a population.}

\item{r}{Intrinsic growth rate.}

\item{K}{Carrying capacity.}
}
\value{
Years until carrying capacity is reached.
}
\description{
Calculate the number of years until a population reaches carrying capacity
}
